+++ src/App.jsx
@@ -1,32 +1,14 @@
 import React from 'react';
 
-import { useDispatch, useSelector } from 'react-redux';
-
-import Page from './Page';
-
-import { updateTaskTitle, addTask } from './actions';
+import InputContainer from './InputContainer';
+import ListContainer from './ListContainer';
 
 export default function App() {
-  const dispatch = useDispatch();
-
-  const { taskTitle } = useSelector((state) => ({
-    taskTitle: state.taskTitle,
-    tasks: state.tasks,
-  }));
-
-  function handleChangeTitle(event) {
-    dispatch(updateTaskTitle(event.target.value));
-  }
-
-  function handleClickAddTask() {
-    dispatch(addTask());
-  }
-
   return (
-    <Page
-      taskTitle={taskTitle}
-      onChangeTitle={handleChangeTitle}
-      onClickAddTask={handleClickAddTask}
-    />
+    <div>
+      <h1>To-do</h1>
+      <InputContainer />
+      <ListContainer />
+    </div>
   );
 }
