+++ src/Page.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { render, fireEvent } from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import { useSelector } from 'react-redux';
 
@@ -15,24 +15,14 @@ test('Page', () => {
   ];
 
   useSelector.mockImplementation((selector) => selector({
+    taskTitle: '',
     tasks,
   }));
 
-  const handleChangeTitle = jest.fn();
-  const handleClickAddTask = jest.fn();
-
   const { getByText } = render((
-    <Page
-      taskTitle=""
-      onChangeTitle={handleChangeTitle}
-      onClickAddTask={handleClickAddTask}
-    />
+    <Page />
   ));
 
   expect(getByText(/Task-1/)).not.toBeNull();
   expect(getByText(/Task-2/)).not.toBeNull();
-
-  fireEvent.click(getByText('추가'));
-
-  expect(handleClickAddTask).toBeCalled();
 });
